HD frame lists and turn analysis (#14)
* adapt correlation analysis to for phase restriction, untested, commit before pulling from master

* commit before pulling

* phase-specific correlation function ready to be tested. create figures/correlation/phaseSpecific folder before testing so figures can be saved accordingly

* add script to plot distribution of turn features

* update analysis to include phase restriction to multiple scripts, including for N2 movies

* add 1W turn figures

* modify turnFeatures.m to compare cluster leaving vs lone worms

* rename plotTurnFeatures.m

* remove 62.2 from datalist

* modify turn analysis and in cluster proportion to include multiple phase restriction options

* add script to plot integrated head angular speed. need to fix up  plotTurnFeatures.m later

* modify plotTurnFeatures script to plot omega turns as read from features file

* add phases specific omega turn freq plots

* add plot cumulative head angle change script and findLeaveClusterWorms function

* fixed bug in leave cluster worm classification. now need to re-run all the analyses that involve assessing leave cluster worms

* change head angle script to plotratwe per second

* modify plotHeadAngleChange script to add trajectory plotting

* fix bug in head angular speed script to separate discontinuous trajectories following leave cluster/lone worm filers

* implement second division for angle changes

* fix plot head angular speed script to include taking angle difference, also wrote several functions and tidied up script using functions

* add test script for angular speed and fixed bugs with regard to smoothing according to test results

* fixed smoothing methods

* before for SK day

* update angular speed script with newly written functions, also fixed small bug in leave cluster detection

* plot three head angle features in parallel, tidy up code

* tidy up script and run it for different dataset and markers

* add analysis results

* update turn analysis figures and remove deprecated analysis results

* Delete .ariadne.config

* Delete .ariadne

* remove hidden files

* remove deprecated pdf files and duplicate .xlsx files

* modify/add script to save/visualise all filtered trajectories regardless of traj length

* modify and rename findLeaveCluster function to include additional worm categories; rename other analysis scripts to start with 'analyze'

* add renamed analysis scripts

* rename scripts to distinguish between visualization and analysis

* rename result figures to be in line with current naming convention regarding dataset and marker

* renaming figures to fit current naming convention

* pull analysis results from SK

* update keynote overall panels for turn analysis

* add phase frames for HD movies to datalist folder

* rotate trajectories for visualisation

* update sample head traj figs after fixing bug in file saving name, working progress in leave cluster filtering data loss

* merge

* add script to calculate leave cluster filter effects

* fiddling with manually joined red traj

* convert eps files

+++ getWormClusterStatus.m
@@ -7,7 +7,7 @@ function [ numCloseNeighbrs, mindist] = getWormClusterStatus(trajData, frame,...
 [x, y] = getWormPositions(trajData, frame, false);
 
 if numel(x)>1 % need at least two worms in frame to calculate distances
-    D = squareform(pdist([x y]).*pixelsize); % distance of every worm to every other
+    D = squareform(pdist([double(x) double(y)]).*double(pixelsize)); % distance of every worm to every other
     % find lone worms
     mindist = min(D + max(max(D))*eye(size(D)));
     % find worms in clusters
